goo/com/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pip==18.0
requests==2.19.1
setuptools==40.2.0
Werkzeug==0.14.1
wheel==0.31.1
google-cloud-error-reporting==0.30.0
google-cloud-datastore
netifaces